+++ inte_edc/settings/debug.py
@@ -5,7 +5,7 @@ from .defaults import *  # noqa
 
 print(f"Settings file {__file__}")  # noqa
 
-SITE_ID = SiteID(default=1)
+SITE_ID = SiteID(default=103)
 EDC_SITES_UAT_DOMAIN = False
 DEBUG = True
 ALLOWED_HOSTS = [