+++ inte_subject/models/clinical_review_baseline.py
@@ -1,5 +1,3 @@
-from django.db import models
-from edc_constants.choices import YES_NO
 from edc_model import models as edc_models
 from edc_visit_schedule.constants import DAY1
 
@@ -9,6 +7,7 @@ from ..model_mixins import (
     ClinicalReviewBaselineHtnModelMixin,
     ClinicalReviewBaselineDmModelMixin,
     ClinicalReviewModelMixin,
+    TreatmentPayMethodsModelMixin,
 )
 
 
@@ -17,6 +16,7 @@ class ClinicalReviewBaselineError(Exception):
 
 
 class ClinicalReviewBaseline(
+    TreatmentPayMethodsModelMixin,
     ClinicalReviewBaselineHivModelMixin,
     ClinicalReviewBaselineHtnModelMixin,
     ClinicalReviewBaselineDmModelMixin,
@@ -24,19 +24,6 @@ class ClinicalReviewBaseline(
     CrfModelMixin,
     edc_models.BaseUuidModel,
 ):
-
-    health_insurance = models.CharField(
-        verbose_name="Does the patient have any private or work-place health insurance?",
-        max_length=15,
-        choices=YES_NO,
-    )
-
-    patient_club = models.CharField(
-        verbose_name="Does the patient belong to a ‘club’ that supports medicines purchase?",
-        max_length=15,
-        choices=YES_NO,
-    )
-
     def save(self, *args, **kwargs):
         if (
             self.subject_visit.visit_code != DAY1