+++ inte_subject/models/htn_initial_review.py
@@ -18,7 +18,7 @@ class HtnInitialReview(
         default=NOT_APPLICABLE,
     )
 
-    med_start_ago = edc_models.DurationYearMonthField(
+    med_start_ago = edc_models.DurationYMDField(
         verbose_name=(
             "If the patient is taking medicines for hypertension, "
             "how long have they been taking these?"