+++ inte_visit_schedule/visit_schedules/schedule.py
@@ -75,7 +75,7 @@ visit00 = Visit(
     rupper=relativedelta(days=0),
     requisitions=requisitions_d1,
     crfs=crfs_d1,
-    facility_name="7-day-clinic",
+    facility_name="5-day-clinic",
 )
 
 
@@ -84,11 +84,11 @@ visit06 = Visit(
     title="Month 6",
     timepoint=6,
     rbase=relativedelta(months=6),
-    rlower=relativedelta(days=0),
-    rupper=relativedelta(days=6),
+    rlower=relativedelta(months=1),
+    rupper=relativedelta(months=5),
     requisitions=requisitions_6m,
     crfs=crfs_6m,
-    facility_name="7-day-clinic",
+    facility_name="5-day-clinic",
 )
 
 visit12 = Visit(
@@ -96,11 +96,11 @@ visit12 = Visit(
     title="Month 12",
     timepoint=12,
     rbase=relativedelta(months=12),
-    rlower=relativedelta(days=0),
-    rupper=relativedelta(days=6),
+    rlower=relativedelta(months=1),
+    rupper=relativedelta(months=3),
     requisitions=requisitions_12m,
     crfs=crfs_12m,
-    facility_name="7-day-clinic",
+    facility_name="5-day-clinic",
 )
 
 schedule_hiv.add_visit(visit=visit00)